Trunk sources reverted to laste released version - 2.8 R3 2.9 moved to Beta section

+++ libraries/AP_Limits/AP_Limit_Geofence.cpp
@@ -144,7 +144,7 @@ AP_Int8 AP_Limit_Geofence::fence_total() {
 // save a fence point
 void AP_Limit_Geofence::set_fence_point_with_index(Vector2l &point, uint8_t i)
 {
-    uintptr_t mem;
+    uint32_t mem;
 
     if (i >= (unsigned)fence_total()) {
         // not allowed
@@ -165,7 +165,7 @@ void AP_Limit_Geofence::set_fence_point_with_index(Vector2l &point, uint8_t i)
  */
 Vector2l AP_Limit_Geofence::get_fence_point_with_index(uint8_t i)
 {
-    uintptr_t mem;
+    uint32_t mem;
     Vector2l ret;
 
     if (i > (unsigned) fence_total()) {
